Updated terminal-related types from int to long across multiple files to accommodate larger terminal IDs. Refactored `WebSocketServerSibs` to introduce modular methods (`RegisterTerminal`, `RegisterFront`) for streamlined terminal registration and linking. Added handling for `SET_AUTH_CREDENTIAL_REQUEST` and improved `HEARTBEAT_NOTIFICATION` logic. Enhanced scalability and maintainability of the codebase.

+++ SimpleWebSocketServer.SIBS.Server.Service/SocketService.cs
@@ -12,7 +12,7 @@ namespace SimpleWebSocketServer.SIBS.Server.Service
         private WebSocketServerSibs _server;
 
         public bool IsStarted => _server.IsStarted;
-        public ConcurrentDictionary<Guid, int> Terminals => _server.Terminals;
+        public ConcurrentDictionary<Guid, long> Terminals => _server.Terminals;
         public ConcurrentDictionary<Guid, Guid> Fronts => _server.Fronts;
         public ConcurrentDictionary<Guid, Guid> TerminalToFrontMap => _server.TerminalToFrontMap;
